admin/requirements.txt
fastapi
uvicorn
httpx
google-analytics-data
google-auth
